requirements.txt
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.4
feedparser==6.0.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pyaes==1.6.1
pyasn1==0.6.2
python-dotenv==1.0.1
python-telegram-bot==21.6
requests==2.32.3
rsa==4.9.1
sgmllib3k==1.0.0
Telethon==1.42.0
typing_extensions==4.15.0
urllib3==2.6.3
